🎨 Refactor permutations for readability
Negligible performance impact.

 * Extracted permutations into their own modules, for ease of review.
 * Added JSDoc documentation for internal utility functions `xor()`, `copy()`, and `rotate()`.
 * Adjusted function signatures for internal utility functions for better readability:
   * `copy(source, offset)(target, offset)`
   * `xor(output, offset)(input1, offset1, ...)`
 * Added `reset()` after permutations to zero out the internal intermediate variables, for security hygiene.
