+++ src/sponge/index.js
@@ -1,47 +1,33 @@
 import { Buffer } from 'buffer';
-import keccak from './keccak';
+import permute from './permute';
 
 const allowedCapacityValues = [224, 256, 384, 512];
 
-const fill = (A, value, offset = 0) => {
-  for (let i = offset; i < A.length; i++) {
-    A[i] = value;
+const writeState = (I, O) => {
+  for (let i = 0; i < I.length; i += 8) {
+    const o = i / 4;
+    O[o] ^= I[i + 7] << 24 | I[i + 6] << 16 | I[i + 5] << 8 | I[i + 4];
+    O[o + 1] ^= I[i + 3] << 24 | I[i + 2] << 16 | I[i + 1] << 8 | I[i];
   }
 };
 
-// eslint-disable-next-line max-params
-const toWord = (I, i, O, o) => {
-  O[o] ^= I[i + 7] << 24 | I[i + 6] << 16 | I[i + 5] << 8 | I[i + 4];
-  O[o + 1] ^= I[i + 3] << 24 | I[i + 2] << 16 | I[i + 1] << 8 | I[i];
-};
+// eslint-disable-next-line max-statements
+const readHash = (I, n) => {
+  const O = Buffer.allocUnsafe(n);
 
-// eslint-disable-next-line max-params
-const toBytes = (I, i, O, o) => {
-  O[o] = I[i + 1];
-  O[o + 1] = I[i + 1] >>> 8;
-  O[o + 2] = I[i + 1] >>> 16;
-  O[o + 3] = I[i + 1] >>> 24;
-  O[o + 4] = I[i];
-  O[o + 5] = I[i] >>> 8;
-  O[o + 6] = I[i] >>> 16;
-  O[o + 7] = I[i] >>> 24;
-};
-
-const commit = (queue, state, permute) => {
-  for (let i = 0; i < queue.length; i += 8) {
-    toWord(queue, i, state, i / 4);
-  }
-  permute(state);
-};
-
-const exportHash = (state, hashSize) => {
-  const hash = Buffer.allocUnsafe(hashSize);
-
-  for (let i = 0; i < hashSize; i += 8) {
-    toBytes(state, i / 4, hash, i);
+  for (let o = 0; o < n; o += 8) {
+    const i = o / 4;
+    O[o] = I[i + 1];
+    O[o + 1] = I[i + 1] >>> 8;
+    O[o + 2] = I[i + 1] >>> 16;
+    O[o + 3] = I[i + 1] >>> 24;
+    O[o + 4] = I[i];
+    O[o + 5] = I[i] >>> 8;
+    O[o + 6] = I[i] >>> 16;
+    O[o + 7] = I[i] >>> 24;
   }
 
-  return hash;
+  return O;
 };
 
 // eslint-disable-next-line max-statements
@@ -50,7 +36,7 @@ const Sponge = function({ capacity, padding }) {
     throw new Error('Unsupported hash length');
   }
 
-  const permute = keccak();
+  const keccak = permute();
 
   const stateSize = 200;
   const hashSize = capacity / 8;
@@ -66,7 +52,8 @@ const Sponge = function({ capacity, padding }) {
       queueOffset += 1;
 
       if (queueOffset >= queueSize) {
-        commit(queue, state, permute);
+        writeState(queue, state);
+        keccak(state);
         queueOffset = 0;
       }
     }
@@ -84,19 +71,20 @@ const Sponge = function({ capacity, padding }) {
       output.state[i] = state[i];
     }
 
-    fill(output.queue, 0, queueOffset);
+    output.queue.fill(0, queueOffset);
 
     output.queue[queueOffset] |= padding;
     output.queue[queueSize - 1] |= 0x80;
 
-    commit(output.queue, output.state, permute);
+    writeState(output.queue, output.state);
+    keccak(output.state);
 
-    return exportHash(output.state, hashSize);
+    return readHash(output.state, hashSize);
   };
 
   this.reset = () => {
-    fill(queue, 0);
-    fill(state, 0);
+    queue.fill(0);
+    state.fill(0);
     queueOffset = 0;
     return this;
   };
