+++ src/sponge/permute/iota/round-constants/index.js
@@ -1,0 +1,28 @@
+const ROUND_CONSTANTS = new Uint32Array([
+  0x00000000, 0x00000001,
+  0x00000000, 0x00008082,
+  0x80000000, 0x0000808a,
+  0x80000000, 0x80008000,
+  0x00000000, 0x0000808b,
+  0x00000000, 0x80000001,
+  0x80000000, 0x80008081,
+  0x80000000, 0x00008009,
+  0x00000000, 0x0000008a,
+  0x00000000, 0x00000088,
+  0x00000000, 0x80008009,
+  0x00000000, 0x8000000a,
+  0x00000000, 0x8000808b,
+  0x80000000, 0x0000008b,
+  0x80000000, 0x00008089,
+  0x80000000, 0x00008003,
+  0x80000000, 0x00008002,
+  0x80000000, 0x00000080,
+  0x00000000, 0x0000800a,
+  0x80000000, 0x8000000a,
+  0x80000000, 0x80008081,
+  0x80000000, 0x00008080,
+  0x00000000, 0x80000001,
+  0x80000000, 0x80008008
+]);
+
+export default ROUND_CONSTANTS;
