+++ src/sponge/permute/rho-pi/pi-shuffles/index.js
@@ -1,0 +1,3 @@
+const PI_SHUFFLES = [10, 7, 11, 17, 18, 3, 5, 16, 8, 21, 24, 4, 15, 23, 19, 13, 12, 2, 20, 14, 22, 9, 6, 1];
+
+export default PI_SHUFFLES;
