+++ src/sponge/permute/rho-pi/rho-offsets/index.js
@@ -1,0 +1,3 @@
+const RHO_OFFSETS = [1, 3, 6, 10, 15, 21, 28, 36, 45, 55, 2, 14, 27, 41, 56, 8, 25, 43, 62, 18, 39, 61, 20, 44];
+
+export default RHO_OFFSETS;
